sample_code/blockchain_and_storage/nebula_block_storage/requirements.txt
boto3==1.34.11
python-dotenv==1.0.0 